refactor: removes duplicate code (recipient forms)
- creates common ActionContainer component for action buttons
- extracts fields for Add/Edit Recipient forms into common component
