+++ src/features/user/BulkCodes/components/ActionContainer.tsx
@@ -1,0 +1,11 @@
+import { Box, styled } from '@mui/material';
+
+const ActionContainer = styled(Box)({
+  height: 40,
+  display: 'inline-flex',
+  justifyContent: 'center',
+  alignItems: 'center',
+  margin: '0 3px',
+});
+
+export default ActionContainer;
